+++ CHANGELOG.md
@@ -3,5 +3,6 @@
 - Added category cms title support
 - Added product uvp support
 - Added seo support for product image filenames
+- Added attribute for product active flag
 - Fixed active flag for product parents to 0
 - Fixed category and manufacturer image key mapping